Add external blueprint for linking to pages from other frontend apps
Templates are using url_for with flask view names to generate links
to other application pages. After the split, some of these pages no
longer exist in the same application.

Normally, this would mean that we need to replace the `url_for` calls
with inline URLs. Instead, we add a new `external` blueprint without
the app's url prefix.

This allows us to copy the view definitions from the buyer-frontend
and keep using `url_for` with the new `external.*` view names.

The views aren't actually called, since the link will be routed to
a different app, but we get all the safety benefits of using `url_for`
instead of hardcoded URLs.

In the future, we could move the external view and blueprint to utils
so it can be shared between the apps.