Preserve line breaks for textarea formatted text
We're going to allow newline formatting for `textbox_large`
content submitted to briefs, as well as for both submitted
questions and answers for briefs.

Here, we're showing the line breaks back to buyers on their
brief editing summary table pages (when they're building a
brief), as well as on the page where all questions and answers
are listed.