Instrument GA tracking to capture clicks made on external links
Removes wrapper function in favour of a direct call to GOVUK function. This is required as wrapper function is not available in all apps.

====
A lot of GA tracking comes 'out of the box' but a notable exception is tracking to capture clicks on external links. Additional code is required in order to capture these.

Currently, we are tracking some external links on the DM site (those that were specifically required for previous mission work).

The proposal is to have all external links tracked automatically across the whole site, similar to how GOV.UK have done.

The advantage of doing this is that we have key data available straight away at the start of mission teams rather than having to prioritise the work, use up dev resource and then wait for enough data to populate before being able to carry out analysis.

---> includes making sure all links on the homepage are tracked, so we can measure which links are used the most and least.

https://trello.com/c/GU1raroG/57-instrument-ga-tracking-to-capture-clicks-made-on-external-links