Rename framework_framework variable
 ## Summary
`framework_framework` is an unnecessarily obtuse variable name and is
not overly descriptive of what it actually means. When used, this
describes the 'family' of framework being referenced. For example,
G-Cloud 10 is in the 'G-Cloud' family of frameworks. Digital Outcomes
and Specialists 2 is in the 'Digital Outcomes and Specialists' family.
Let's change all the references of `framework_framework` to
`framework_family`.

 ## Ticket
https://trello.com/c/iqFcVKpd/14

 ## Dependencies
https://github.com/alphagov/digitalmarketplace-utils/pull/397