Refactor validation masthead use
 ## Summary
We have an updated version of the validation masthead for pages
displaying errors with form submissions that abstracts some more of the
responsibility from the individual templates and supports our main error
formats natively (content loader forms and wtforms). This means we can
rip out some of the boilerplate from our templates.

There's also a page here that didn't use the validation masthead from
the frontend-toolkit which has now been brought in line.

 ## Ticket
https://trello.com/c/RSYEM3FL/448