remove use of FEATURE_FLAGS_NEW_SUPPLIER_FLOW, replace with is_legacy_brief_response logic
we want to retire the use of flask feature flags and this is one of the last
uses of them. previously we were comparing the `publishedAt` date with
a feature flag's `enabled_since` date to determine whether it's a "legacy brief"
or not. instead use the logic of `is_legacy_brief_response`, adapted from
brief-responses-frontend, to figure this out. this function works by
inspecting a brief *response*, so doesn't work in cases where there are no
responses. fortunately the only bit of template that want to know its
legacy status make no sense if there aren't any brief responses, so that
largely solves that. update tests to provide information needed by
`is_legacy_brief_response` in mocks.