Load supplier lot csvs from correct framework s3 bucket
The csvs of suppliers by lot were being loaded from the origin DOS
framework bucket loaction, regardless of the framework. This is weird
and confusing and this fixes that.

The files will need copying over to the correct locations to make sure
this doesn't 404. That will be done before this is released.