Rewrite tests for requirements task list page
I've tried to make the tests for this view easier to follow and simpler
to maintain. I've

- split tests into different classes by brief status
- added a mixin class for sharing fixtures
- pulled out tests of things that should be true no matter what the
brief status
- used snapshot testing for the task list HTML instead of testing links,
as I found those tests were hard to follow
- put tests of "sad paths" into a separate test class

I've also added a bunch of tests for the task list page pattern in the
draft brief state, when all the tags are present.