Fix preview_brief_source
alphagov/digitalmarketplace-content-loader#126 made the
`summary_page_description` field a template field, and
alphagov/digitalmarketplace-frameworks#665 made this field vary by the
lot slug. The brief view in the buyer frontend filters the
`display_brief` manifest by the brief[[1]], so the template context
variable for the lot slug is `lotSlug`, not `lot`.

This commit changes the `preview_brief_source` view to do the same as
the buyer frontend does and pass the whole brief to the filter context.
The buyer frontend relies on `lot` being in the brief object, however
this is deprecated[[2]], so in addition we add `lot` in just in case.

[1]: https://github.com/alphagov/digitalmarketplace-buyer-frontend/blob/a54b9a071525ddcacac071973d3be65ed96c2b0d/app/main/views/marketplace.py#L183
[2]: https://github.com/alphagov/digitalmarketplace-api/blob/a09e09dceb7d0eb995cac6ec914cbaf4d611be1c/app/models/main.py#L1735