prevent 500 when viewing timeline after publishing brief
Previously we didn't handle the case where the buyer viewed the timeline after publishing an opportunity with a question and answer session.
The intended behaviour in this case, as confirmed by @NoraGDS is that they can still see the question and answer session on the timeline after publishing (but without the Edit link). This is what is implemented by this PR.